Added new Single and Binary Fiitting Files
These updated files use the square root weight system to find the best model with the least chi squared value, then uses the other system of ones to refine the flux and do the residual error - this system turned out having the lowest errors compared to the others. Next step will be to have python find the index of the max pixel value, but also make sure that it is in fact the star and not a cosmic ray or something.