Improved MonteCarlo
This revised code only does the montecarlo random noise stuff to the top three fits/models, which significantly improves the runtime. Additionally, the code now deletes the files it creates when it is done. The code still needs to place the output files in a separate directory within the folder with the cpp files, and the variables for selecting the best 3 fits should be modified so that the number of fits it runs through can be changed.